video compression

+++ src/cron/index.cron.ts
@@ -4,22 +4,21 @@ import { queueTable } from "db/schema";
 import { and, eq, sql } from "drizzle-orm";
 import { unlink } from "fs/promises";
 import { CleanPromise } from "utils/cleanPromise";
+async function CompressionsJob() {
+  const twoHoursAgo = new Date(Date.now() - 60 * 60 * 1000 * 2); // 2 hours in milliseconds
 
-export const CheckForOldCompressions = new CronJob(
-  "*/60 * * * *", // cronTime
-  async function () {
-    const twoHoursAgo = new Date(Date.now() - 60 * 60 * 1000 * 2); // 2 hours in milliseconds
+  const db = await Initdatabase();
+  const oldData = await db
+    .select()
+    .from(queueTable)
+    .where(
+      and(
+        eq(queueTable.status, "Completed"),
+        sql`${queueTable.completedAt} > ${twoHoursAgo}`
+      )
+    );
 
-    const db = await Initdatabase();
-    const oldData = await db
-      .select()
-      .from(queueTable)
-      .where(
-        and(
-          eq(queueTable.status, "Completed"),
-          sql`${queueTable.completedAt} < ${twoHoursAgo}`
-        )
-      );
+  if (oldData) {
     for (let index = 0; index < oldData.length; index++) {
       await CleanPromise(unlink(`temp/${oldData[index].compressedName}`));
     }
@@ -31,7 +30,12 @@ export const CheckForOldCompressions = new CronJob(
           sql`${queueTable.createdAt} < ${twoHoursAgo}`
         )
       );
-  }, // sch
+  }
+}
+CompressionsJob();
+export const DeleteOldCompressions = new CronJob(
+  "*/60 * * * *", // cronTime
+  CompressionsJob,
   null, // onComplete
   false // start
 );
